Move styleConstants to static dir
