+++ src/app/components/LinearProgress/styles.js
@@ -1,6 +1,4 @@
-import { styleConstants } from '../../config';
-
-const { colors } = styleConstants;
+import { colors } from '../../static/styleConstants';
 
 const styles = {
   container: {
