+++ src/app/components/Version/styles.js
@@ -1,8 +1,6 @@
 import css from 'styled-jsx/css'; // eslint-disable-line
 
-import { styleConstants } from '../../config';
-
-const { rhythm } = styleConstants;
+import { rhythm } from '../../static/styleConstants';
 
 const styles = css`
   .container {
