+++ src/app/static/styleConstants/index.js
