+++ src/app/static/styles/global/index.js
@@ -1,8 +1,6 @@
 import css from 'styled-jsx/css';
 
-import { styleConstants } from '../../../config';
-
-const { colors } = styleConstants;
+import { colors } from '../../styleConstants';
 
 const styles = css.global`
   * {
